backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
arq>=0.26.0
pydantic
pydantic-settings
qdrant-client>=1.6.4
supabase>=2.3.0
openai
anthropic
python-dotenv==1.0.0
python-multipart==0.0.6
PyPDF2==3.0.1
pymupdf4llm
pymupdf
markdownify
langgraph
langchain-openai
langchain-core
reportlab
python-docx>=1.0.0
weasyprint>=60.0
slowapi>=0.1.9
redis>=5.0.0
tiktoken>=0.7.0
slowapi
